# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/user/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Details",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    }
                }
            }
        },
        "/api/user/{id}/certification": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get certification by user id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Certification Details",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    }
                }
            }
        },
        "/api/user/{id}/education": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get Education by user id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Education Details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.BaseResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "output_schema": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.UserEducationResp"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "educations": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/dto.Education"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    }
                }
            }
        },
        "/api/user/{id}/personal-information": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get personal information by user id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Personal Information Details",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    }
                }
            }
        },
        "/api/user/{id}/skill": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get skill by user id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Skill Details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.BaseResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "output_schema": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.UserSkillResp"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "skills": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/dto.Skill"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    }
                }
            }
        },
        "/api/user/{id}/work-experience": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get work experience by user id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Work Experience Details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.BaseResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "output_schema": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.UserWorkExperienceResp"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "work_experiences": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/dto.WorkExperience"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorSchema"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.BaseResp": {
            "type": "object",
            "properties": {
                "error_schema": {
                    "$ref": "#/definitions/dto.ErrorSchema"
                },
                "output_schema": {
                    "description": "OutputSchema fleksibel (bisa berisi apa saja)"
                }
            }
        },
        "dto.Education": {
            "type": "object",
            "properties": {
                "degree": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "education_id": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "field_of_study": {
                    "type": "string"
                },
                "school_name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorMessage": {
            "type": "object",
            "properties": {
                "en": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorSchema": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string"
                },
                "error_message": {
                    "$ref": "#/definitions/dto.ErrorMessage"
                }
            }
        },
...
        "dto.Skill": {
            "type": "object",
            "properties": {
                "level": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "skill_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.UserEducationResp": {
            "type": "object",
            "properties": {
                "educations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Education"
                    }
                }
            }
        },
...
        "dto.UserSkillResp": {
            "type": "object",
            "properties": {
                "skills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Skill"
                    }
                }
            }
        },
        "dto.UserWorkExperienceResp": {
            "type": "object",
            "properties": {
                "work_experiences": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.WorkExperience"
                    }
                }
            }
        },
        "dto.WorkExperience": {
            "type": "object",
            "properties": {
                "company_name": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "job_description": {
                    "type": "string"
                },
                "job_title": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "work_experience_id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}